Extension now works again, removed `browser-polyfill`, and split the extension into two parts
The extension now functions relatively well on both Chrome and Firefox, at the expense of spliting the extension into two, distinct versions for each browser.

Firefox does not officially support Manifest V3 (in that it does not support the new `service_worker` key, among others), which Chrome requires for any new extensions being submitted to the Chrome Web Store.

The `webextension-polyfill` library is unable to rectify this issue, and I encountered a few issues when trying to import the library into background and content scripts manually, which is why it has been removed and the extension split into two separate versions - one for Chrome, and one for Firefox. (At least until Firefox implements support for Manifest V3.)